etl/requirements.txt
pandas
openpyxl
boto3
python-dotenv
pytest
moto
pyarrow
localstack-client
loguru
